sql: rename transform module to transform_ast
This better reflects its purpose, and makes space for a separate module
that will transform the IR.
